slides/20221015pyconjp/code/requirements.txt
deepl==1.11.0
google-api-python-client==2.64.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
jira==3.4.1
peewee==3.15.3
requests==2.28.1
slack-bolt==1.15.1
slack-sdk==3.19.1
sympy==1.11.1
